[sycl] moved PluginWrapper in cms::sycltools
