+++ src/sycl/SYCLCore/Product.h
@@ -57,14 +57,5 @@ namespace cms {
 
   }  // namespace sycltools
 }  // namespace cms
-template <typename T, typename P>
-class PluginWrapper {
-public:
-  template <typename... Args>
-  explicit PluginWrapper(Args&&... args) : obj_{std::forward<Args>(args)...} {}
-  T const& get() const { return obj_; }
-private:
-  T obj_;
-};
 
 #endif
